Update method "show" in MealController
