+++ app/Http/Requests/Favorites/AddFavoriteRequest.php
@@ -2,26 +2,9 @@
 
 namespace App\Http\Requests\Favorites;
 
-use Illuminate\Foundation\Http\FormRequest;
+use App\Http\Requests\Meals\GetMealByIdRequest;
 
-class AddFavoriteRequest extends FormRequest
+class AddFavoriteRequest extends GetMealByIdRequest
 {
-    public function authorize()
-    {
-        return true;
-    }
-
-    public function rules()
-    {
-        return [
-            'meal_id' => 'required|integer|exists:meals,id',
-        ];
-    }
-
-    public function all($keys = null)
-    {
-        $data = parent::all();
-        $data['meal_id'] = $this->route('id');
-        return $data;
-    }
+ //
 }
