+++ app/Repositories/MealsRepository/MealsRepositoryInterface.php
@@ -5,6 +5,7 @@ namespace App\Repositories\MealsRepository;
 interface MealsRepositoryInterface
 {
     public function getAll(): array;
+    public function getById(int $mealId): array;
     public function getByIngredients(array $idx): array;
     public function getByIngredientId(int $id): array;
     public function getRandom(): array;
